Map editor (#40)
* WIP on the text editor.

* Yayayayay

* Enumbased editor

* Borko

* Added a mappymappymap

* Map editor command map patterns.

* Adds helper to start map editor.

* let the ai generate a little map for testing.

* Fixes player spawning and map formatting.

* Created a copy of the map

* Fixes spawning issue.

* Updated... a lot of stuff?

* Tweaking the map editor

* Modified startup map

* Tweaking the map editor to make it TIGHT

* Updated map after updating map editor.

* Changed some stuff regards to player.

* There we go, now it works, perhaps maybe.

* Whateeevs

* Maybe fixed?

* The status of the wall stuff.

* Now the things are right.

* Wall things work as intended now. Map editor is go.

* Yay, this is pleasing.

* Very cool.

* Pixelate shader added?

* OK; So the map editor in fact works, right now, which is cool, I guess.

* OK, so map editor prototype done, I guess. Read the notes again!

+++ src/general/components/map_components.rs
@@ -58,9 +58,10 @@ pub struct MapFile {
     pub map_width: usize,
     #[serde(default = "default_map_height")]
     pub map_height: usize,
-    /// Row-major grid; 0=void, 1=floor, 5=alien spawn, 9=alien goal, 17=player spawn.
+    /// Row-major grid of `BitFlags<MapFeatures>` values stored as u64.
+    /// 0 = void (Nothing), non-zero = tile with feature flags set.
     #[serde(default)]
-    pub tiles: Vec<Vec<u8>>,
+    pub tiles: Vec<Vec<u64>>,
     #[serde(default)]
     pub decorations: Vec<DecorationItem>,
     /// Items placed by the map editor (def-driven models on specific tiles).
